refactor(jobs): collapse JobsOperations/JobsStatus passthrough into Jobs
Closes #1787.

The Jobs class was an 18/19 passthrough facade delegating to JobsOperations
and JobsStatus. With 29 external callers using `new Jobs(...)`, the cleanest
fix is to absorb the satellites back into Jobs rather than push delegation
out to every call site.

Changes:
- Jobs now extends BaseRepository directly and owns CRUD, status transitions,
  flow health bookkeeping, and schema migrations in one place.
- JobsOperations and JobsStatus deleted.
- JobCompleteHandler switched from `new JobsOperations()` to `new Jobs()`
  (update_flow_health_cache and friends are now first-class on Jobs).
- Removed an unused `use ... JobsOperations` import in inc/Api/System/System.php.
- PostTrackingTest fixed: previously did `new JobsOperations($jobs_db)` against
  a 0-arg constructor (PHP swallowed the extra arg). Now uses Jobs directly.
- Updated docblock references in tests/jobs-get-children-smoke.php.

Verified:
- `php -l` clean on all touched files.
- `composer dump-autoload -o` regenerates clean (1033 classes).
- `php tests/jobs-get-children-smoke.php` → all pass.
- `homeboy audit` shows facade_passthrough findings dropped from 1 → 0.

No behavior change. External `new Jobs()` API preserved.
